pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "apple-subscription-validator"
version = "2.0.0"
description = "Python library and CLI for validating Apple In-App Purchase receipts and transactions"
readme = "README.md"
requires-python = ">=3.7"
license = {text = "MIT"}
keywords = ["apple", "app-store", "subscription", "receipt", "validation", "jws", "storekit", "iap"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Utilities",
]

dependencies = [
    "PyJWT[crypto]==2.8.0",
    "cryptography==41.0.7",
    "requests==2.31.0",
    "python-dotenv==1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[project.scripts]
apple-sub-validator = "apple_subscription_validator:main"
apple-sub-validator-interactive = "interactive_validator:interactive_validate"

[tool.setuptools]
py-modules = ["apple_subscription_validator", "interactive_validator", "validate_from_file"]

[tool.black]
line-length = 120
target-version = ['py37', 'py38', 'py39', 'py310', 'py311']

[tool.ruff]
line-length = 120
target-version = "py37"

[tool.mypy]
python_version = "3.7"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
